Fix states: edit buttons logic
