+++ tgbot/handlers/create_poll.py
@@ -4,7 +4,7 @@ from aiogram.types import Message, CallbackQuery
 from aiogram.fsm.state import StatesGroup, State
 from aiogram.fsm.context import FSMContext
 
-import keyboards.check as kb
+import keyboards.create_poll as kb
 
 router = Router()
 
@@ -13,7 +13,6 @@ class CreatePoll(StatesGroup):
     name = State()
     question = State()
     answer = State()
-    correct = State()
 
 
 # Обработка нажатия кнопки (from start) и переход в состояние
@@ -37,7 +36,7 @@ async def process_name(message: Message, state: FSMContext):
 async def process_question(message: Message, state: FSMContext):
     await state.update_data(question=message.text)
     await state.set_state(CreatePoll.answer)
-    await message.answer("Введите ответ на свой вопрос (для подсчетов результатов):")
+    await message.answer("Введите ответ на свой вопрос (для результатов после опроса):")
 
 
 @router.message(CreatePoll.answer)
@@ -53,8 +52,11 @@ async def process_answer(message: Message, state: FSMContext):
 # Обработка нажатия кнопки (from answer) и переход в состояние
 @router.callback_query(F.data == "check_false")
 async def process_check_false(callback: CallbackQuery, state: FSMContext):
+    await callback.message.edit_text("Данные не верны.")
     await callback.answer("Сначала")  # alert
-    await callback.message.answer("Вы начали создание опроса!")  # message in chat
+    await callback.message.answer(
+        "Вы начали создание опроса сначала!"
+    )  # message in chat
 
     await state.set_state(CreatePoll.name)  # goes to state
     await callback.message.answer("Введите название опроса:")
@@ -64,17 +66,9 @@ async def process_check_false(callback: CallbackQuery, state: FSMContext):
 @router.callback_query(F.data == "check_true")
 async def process_check_true(callback: CallbackQuery, state: FSMContext):
     await callback.answer("Успешно")  # alert
-    await callback.message.answer(
-        "Данные верны.", reply_markup=kb.save_menu
-    )  # message in chat
-    await state.set_state(CreatePoll.correct)  # goes to state
-
-
-@router.message(CreatePoll.correct)
-async def process_correct(
-    message: Message, state: FSMContext
-):  # waiting for message need to do a reply
+    await callback.message.edit_text("Данные верны.")
     data = await state.get_data()  # TODO: need to save to db
-    await message.answer("Данные опроса сохранены!")
-    await message.answer(f"Опрос  успешно создан!")
+    await callback.message.answer("Данные опроса сохранены!")
+    await callback.message.answer(f"Опрос {data['name']} успешно создан!")
+
     await state.clear()
