+++ tgbot/handlers/help.py
@@ -1,0 +1,13 @@
+from aiogram import Router, F
+from aiogram.filters import Command
+from aiogram.types import Message
+
+router = Router()
+
+help_text = """Список команд:\n/start - инициализация бота.\n/help - получение информации о боте.\n"""
+
+
+@router.message(Command("help"))
+async def cmd_help(message: Message):
+    """Answer help message with info from `help_text`"""
+    await message.answer(help_text)
